refactor: 내가 쓴 글/댓글 확인 기능 MyActivityService로 이름 변경 -BoardService는 적절한 이름이 아닌것 같아 수정
